Introduce Requester for HTTP Client
