+++ src/Genius/Resources/WebPagesResource.php
@@ -10,7 +10,7 @@ class WebPagesResource extends AbstractResource
 {
     public function get(string $raw_annotatable_url, ?string $canonical_url = null, ?string $og_url = null): \stdClass
     {
-        return $this->getMethod(
+        return $this->requester->get(
             'web_pages/lookup',
             [
                 'raw_annotatable_url' => $raw_annotatable_url,
